+++ build.sbt
@@ -20,7 +20,7 @@ ThisBuild / crossScalaVersions := Seq(Scala213, "3.1.2")
 ThisBuild / scalaVersion := Scala213 // the default Scala
 
 lazy val root = tlCrossRootProject
-  .aggregate(core, java)
+  .aggregate(core, testkit, java)
   .settings(name := "otel4s")
 
 lazy val core = crossProject(JVMPlatform, JSPlatform)
@@ -36,6 +36,21 @@ lazy val core = crossProject(JVMPlatform, JSPlatform)
     )
   )
 
+lazy val testkit = crossProject(JVMPlatform)
+  .crossType(CrossType.Full)
+  .in(file("testkit"))
+  .settings(
+    name := "otel4s-testkit"
+  )
+  .jvmSettings(
+    libraryDependencies ++= Seq(
+      "io.opentelemetry" % "opentelemetry-api" % "1.15.0",
+      "io.opentelemetry" % "opentelemetry-sdk" % "1.15.0",
+      "io.opentelemetry" % "opentelemetry-sdk-testing" % "1.15.0"
+    )
+  )
+  .dependsOn(core)
+
 lazy val java = crossProject(JVMPlatform)
   .crossType(CrossType.Pure)
   .in(file("java"))
@@ -49,6 +64,6 @@ lazy val java = crossProject(JVMPlatform)
       "org.typelevel" %% "munit-cats-effect-3" % "1.0.7" % Test
     )
   )
-  .dependsOn(core % "compile->compile,test->test")
+  .dependsOn(core, testkit)
 
 lazy val docs = project.in(file("site")).enablePlugins(TypelevelSitePlugin)
