+++ testkit/shared/src/main/scala/org/typelevel/otel4s/testkit/Metric.scala
@@ -1,0 +1,8 @@
+package org.typelevel.otel4s.testkit
+
+final case class Metric(
+    name: String,
+    description: Option[String],
+    unit: Option[String],
+    data: MetricData
+)
